[Step5 - 자동차 경주 리팩토링] refactor: Controller의 책임 최소화
