+++ src/main/java/step5/controller/RacingGameController.java
@@ -1,0 +1,48 @@
+package step5.controller;
+
+import step5.Exception.WinnerException;
+import step5.domain.*;
+import step5.view.*;
+
+import java.util.function.Consumer;
+import java.util.stream.Stream;
+
+import static step5.view.ResultView.*;
+
+/**
+ * 초간단 자동차 경주 게임을 구현한다.
+ * 주어진 횟수 동안 n대의 자동차는 전진 또는 멈출 수 있다.
+ * 사용자는 몇 대의 자동차로 몇 번의 이동을 할 것인지를 입력할 수 있어야 한다.
+ * 전진하는 조건은 0에서 9 사이에서 random 값을 구한 후 random 값이 4이상일 경우이다.
+ * 자동차의 상태를 화면에 출력한다. 어느 시점에 출력할 것인지에 대한 제약은 없다.
+ */
+public class RacingGameController {
+
+  final private Racing racing;
+
+  public RacingGameController(MoveStrategy moveStrategy, String[] cars, int time) {
+    this.racing = Racing.of(Cars.of(cars), time, moveStrategy);
+  }
+
+  public void viewRace (Consumer<Stream<Car>> viewer) {
+    while (!racing.isRaceEnd()) {
+      viewer.accept(racing.race());
+    }
+  }
+
+  public Stream<Car> getWinners () {
+    return racing.getWinners();
+  }
+
+  public static void main(String[] args) throws WinnerException {
+    String[] cars = InputView.inputCars();
+    int time = InputView.inputTime();
+    MoveStrategy moveStrategy = RandomNumberMoveStrategy.getInstance();
+    RacingGameController controller = new RacingGameController(moveStrategy, cars, time);
+
+    printResultText();
+    controller.viewRace(ResultView::printRace);
+    printWinner(controller.getWinners());
+  }
+
+}
